% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{missyelliot}
\alias{missyelliot}
\title{Reverse complement}
\usage{
missyelliot(x)
}
\arguments{
\item{x}{A DNA sequence consisting of A, C, G, and T}
}
\value{
A reverse-complemented DNA sequence
}
\description{
Take that flip it and reverse it
}
\examples{
missyelliot("GATTACA")
}
